+++ src/main/java/roomescape/normal/NormalRoomescapeController.java
@@ -1,18 +1,13 @@
-package roomescape;
+package roomescape.normal;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class RoomescapeController {
+public class NormalRoomescapeController {
 
     @GetMapping("/")
     public String getDefaultPage() {
         return "main/index";
-    }
-
-    @GetMapping("/admin")
-    public String getAdminPage() {
-        return "admin/index";
     }
 }
